Got everything working
