+++ apps/mobile/src/screens/bluetooth-device-details-screen.tsx
@@ -1,57 +1,49 @@
-import { useAsync } from "react-async-hook";
-import { Text, View } from "react-native";
+import { useAsync, useAsyncCallback } from "react-async-hook";
+import { View } from "react-native";
 
 import { BigActionButton } from "../components/big-action-button";
 import { ErrorView } from "../components/error-view";
 import { LargeSpinner } from "../components/large-spinner";
-import {
-  bluetoothDeviceToMultispeqStream,
-  connectWithBluetoothDevice,
-} from "../services/bluetooth-classic";
+import { ResultView } from "../components/result-view";
+import { bluetoothDeviceToMultispeqStream } from "../services/multispeq-communication/android-bluetooth-connection/bluetooth-device-to-multispeq-stream";
+import { connectWithBluetoothDevice } from "../services/multispeq-communication/android-bluetooth-connection/connect-with-bluetooth-device";
+import { MultiSpeqCommandExecutor } from "../services/multispeq-communication/multispeq-command-executor";
 
 const protocol = [{ spad: [1] }];
 
 export function BluetoothDeviceDetailsScreen({ route }: any) {
   const { deviceId } = route.params;
 
-  const { result, loading, error } = useAsync(async () => {
+  const {
+    result: multispeq,
+    loading: isConnecting,
+    error,
+  } = useAsync(async () => {
     const device = await connectWithBluetoothDevice(deviceId);
-    const multispeqEmitter = bluetoothDeviceToMultispeqStream(device);
-
-    multispeqEmitter.on("receivedReplyFromDevice", ({ data }) => {
-      console.log("got reply from device", data);
-    });
-
-    multispeqEmitter.emit("sendCommandToDevice", "hello");
-    return { emitter: multispeqEmitter, device };
+    return new MultiSpeqCommandExecutor(
+      bluetoothDeviceToMultispeqStream(device),
+    );
   }, [deviceId]);
 
-  if (loading) {
+  const {
+    execute: handleScan,
+    loading: isScanning,
+    result: scanResult,
+  } = useAsyncCallback(() => multispeq?.execute(protocol));
+
+  if (isConnecting) {
     return <LargeSpinner>Connecting to device...</LargeSpinner>;
   }
 
-  if (error || !result) {
+  if (error || !multispeq) {
     return <ErrorView error={error ?? "Cannot connect"} />;
   }
 
-  const { device, emitter } = result;
-
-  function handleScan() {
-    emitter.emit("sendCommandToDevice", protocol);
-  }
-
   return (
-    <View className="flex-1 p-4 bg-white">
-      <Text className="text-2xl font-bold mb-4">
-        {device.name || "Unknown Device"}
-      </Text>
-      <Text className="text-base text-gray-700 mb-2">ID: {device.id}</Text>
-      <Text className="text-base text-gray-700 mb-2">
-        RSSI: {(device.extra as any)?.rssi ?? "N/A"} dBm
-      </Text>
-      <Text className="text-base text-gray-700">
-        Bonded: {device.bonded ? "Yes" : "No"}
-      </Text>
+    <View className="flex-1 bg-white p-4 justify-between w-full">
+      <View className="flex-[2] w-full justify-center items-center border border-gray-300 rounded-2xl p-4 bg-gray-50 shadow-md">
+        <ResultView scanResult={scanResult} isScanning={isScanning} />
+      </View>
       <BigActionButton onPress={handleScan} text="Start Scan" />
     </View>
   );
