+++ apps/mobile/src/services/multispeq-communication/android-bluetooth-connection/get-bluetooth-classic-devices.ts
@@ -1,0 +1,14 @@
+import RNBluetoothClassic from "react-native-bluetooth-classic";
+
+import { requestBluetoothPermission } from "../../request-bluetooth-permissions";
+
+export async function getBluetoothClassicDevices() {
+  await requestBluetoothPermission();
+  try {
+    await RNBluetoothClassic.cancelDiscovery();
+  } catch (e) {
+    // ignored
+  }
+  const devices = await RNBluetoothClassic.startDiscovery();
+  return devices;
+}
