+++ apps/mobile/src/services/multispeq-communication/android-serial-port-connection/serial-port-events.ts
