+++ apps/mobile/src/services/multispeq-communication/multispeq-stream-events.ts
@@ -1,0 +1,5 @@
+export type MultispeqStreamEvents = {
+  sendCommandToDevice: string | object;
+  receivedReplyFromDevice: { data: object | string; checksum: string };
+  destroy: void;
+};
