+++ apps/mobile/src/utils/stringify-if-object.ts
@@ -1,0 +1,3 @@
+export function stringifyIfObject(value: string | object) {
+  return typeof value === "object" ? JSON.stringify(value) : value;
+}
